refactor($prepare): enhance prepare's structure
